make cin/cout handler common
