+++ src/qrisp/alg_primitives/arithmetic/adders/adder_utilities.py
@@ -22,6 +22,7 @@ import numpy as np
 from qrisp.circuit import Qubit
 from qrisp.core import QuantumVariable
 from qrisp.jasp import DynamicQubitArray, check_for_tracing_mode
+from qrisp.qtypes import QuantumBool
 
 
 def _extract_bit(a_int, digit_index):
@@ -90,3 +91,41 @@ def _validate_adder_inputs(a, b):
             "(classical a, quantum b) or quantum-quantum (quantum a, quantum b)."
         )
     return a_is_quantum, b_is_quantum
+
+
+def _resolve_c_in(c_in):
+    """Resolve the carry-in qubit.
+
+    The carry-in may be passed as a QuantumBool or a bare Qubit. This helper
+    normalizes it to a Qubit, raising a TypeError for any other type in static
+    mode.
+    """
+    if c_in is None:
+        return None
+
+    if isinstance(c_in, QuantumBool):
+        return c_in[0]
+
+    if not check_for_tracing_mode() and not isinstance(c_in, Qubit):
+        raise TypeError(f"c_in must be of type QuantumBool or Qubit, not {type(c_in)}")
+
+    return c_in
+
+
+def _resolve_c_out(c_out):
+    """Resolve the carry-out qubit.
+
+    The carry-out may be passed as a QuantumBool or a bare Qubit. This helper
+    normalizes it to a Qubit, raising a TypeError for any other type in static
+    mode.
+    """
+    if c_out is None:
+        return None
+
+    if isinstance(c_out, QuantumBool):
+        return c_out[0]
+
+    if not check_for_tracing_mode() and not isinstance(c_out, Qubit):
+        raise TypeError(f"c_out must be of type QuantumBool or Qubit, not {type(c_out)}")
+
+    return c_out
